extract Zipper::Zip into it's own file
